+++ model/genobject.py
@@ -1,4 +1,0 @@
-class GenObject:
-    """
-    The base object for the code generator
-    """