+++ model/participant.py
@@ -1,14 +1,16 @@
+from generator.participant import GenParticipant
 from model.simobject import SeObject
 from utils.log_loggers import MODEL_LOG
 
 
-class Participant(SeObject):
+class Participant(SeObject, GenParticipant):
     """
     A general participant in the Smart Energy Simulation
     """
 
     def __init__(self, el_role, en_title, de_title):
-        super().__init__()
+        SeObject.__init__(self)
+        GenParticipant.__init__(self)
         self._el_role = el_role
         self._en_title = en_title
         self._de_title = de_title
@@ -17,3 +19,7 @@ class Participant(SeObject):
     @property
     def el_role(self):
         return self._el_role
+
+    @property
+    def en_title(self):
+        return self._en_title