+++ model/simobject.py
@@ -1,5 +1,7 @@
 import uuid
 
+from utils.log_loggers import MODEL_LOG
+
 
 class SeObject:
     """
@@ -8,3 +10,8 @@ class SeObject:
 
     def __init__(self):
         self._uuid = uuid.uuid4()
+        MODEL_LOG.debug(f"create base object (uuid:'{self._uuid}')")
+
+    @property
+    def uuid(self):
+        return self._uuid