+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
 - mini docker env added
-- new CHANGELOG (see CHANGELOG.md ;-)
+- new CHANGELOG (see [CHANGELOG.md](https://github.com/smart-energy-di/pg-se-sim/blob/main/CHANGELOG.md) ;-)
 - Integrate first test methods
 - Class participant is now also tested
 - each object now has a unique, random UUID
 - new doctests, first Participant doctest
 - Explanation: Why simulation within the "Smart Energy" topic?
+- [doctest](https://github.com/smart-energy-di/pg-se-sim/blob/main/test/test_doc_participant.md) for a first generator run